==> 03_Hindcast_IHData_GOW.ipynb <==
# Obtener información y datos historicos de IHData GOW 
### Para ello, se utilizará el OpenDAP de IHData

import os
import glob
import xarray as xr
import pandas as pd
import folium
import matplotlib.pyplot as plt
import datetime as dt
from tqdm.notebook import tqdm

#### Buscar nodos mas cercanos a las boyas

df_boyas = pd.read_csv('boyas.csv')
df_boyas

def sort_by_year_month(file_names):
    # Función para extraer año y mes de cada nombre
    def get_year_month(file_name):
        parts = file_name.split('_')[-1]  # Obtenemos la parte después del primer '_'
        year, month = parts.split('.')[0], parts.split('.')[1]
        return dt.datetime.strptime(f"{year}-{month}", "%Y-%m")

    # Ordenar usando la función get_year_month como clave
    return sorted(file_names, key=get_year_month)

path = r'\\193.144.213.171\ihcantabria\GOW\GOW_CFS\Europa'
files = glob.glob(os.path.join(path, 'GOW_CFS_Europa_*.nc'))
files = [f for f in files if 'Mask' not in f]
files = sort_by_year_month(files)
for nombre, lat, lon in zip(df_boyas.Nombre[:-2], df_boyas.Latitud[:-2], df_boyas.Longitud[:-2]):
    print(nombre, lat, lon)
    data_res = xr.Dataset()
    for i, f in tqdm(enumerate(files), total=len(files)):
        data = xr.open_dataset(os.path.join(path, f))
        data_c = data.sel(latitude=lat, longitude=lon, method="nearest")     
        if i == 0:
            data_res = data_c
        else:
            data_res = xr.concat([data_res, data_c], dim="time")
    data_res.to_netcdf(f'data/raw/historico/GOW_CFS_{nombre}.nc')

path = r'\\193.144.213.171\ihcantabria\GOW\GOW_CFS\Canarias'
files = glob.glob(os.path.join(path, 'GOW_CFS_Canarias_*.nc'))
files = [f for f in files if 'Mask' not in f]
files = sort_by_year_month(files)
for nombre, lat, lon in zip(df_boyas.Nombre[-2:], df_boyas.Latitud[-2:], df_boyas.Longitud[-2:]):
    print(nombre, lat, lon)
    data_res = xr.Dataset()
    for i, f in tqdm(enumerate(files), total=len(files)):
        data = xr.open_dataset(os.path.join(path, f))
        data_c = data.sel(latitude=lat, longitude=lon, method="nearest")     
        if i == 0:
            data_res = data_c
        else:
            data_res = xr.concat([data_res, data_c], dim="time")
    data_res.to_netcdf(f'data/raw/historico/GOW_CFS_{nombre}.nc')

df_hindcast = pd.DataFrame(columns=['Nombre', 'Longitud', 'Latitud', 'InicioMedida', 'FinMedida'])
for nombre, lat, lon in zip(df_boyas.Nombre, df_boyas.Latitud, df_boyas.Longitud):  
    ds = xr.open_dataset(os.path.join(f'data/raw/historico/GOW_CFS_{nombre}.nc'))
    df_hindcast.loc[len(df_hindcast.index)] = [nombre, ds.longitude.values, ds.latitude.values, ds.time.values[0], ds.time.values[-1]]
df_hindcast

m = folium.Map(location=(39, -5), zoom_start=5)
for i in range(len(df_hindcast)):
    folium.Marker(location=[df_boyas.Latitud[i], df_boyas.Longitud[i]], icon=folium.Icon(color='black')).add_to(m)
    
    folium.Marker(location=[df_hindcast.Latitud[i], df_hindcast.Longitud[i]], tooltip=folium.Tooltip(df_hindcast.Nombre[i], permanent=True)).add_to(m)
m

fig, ax = plt.subplots()
for i, task in df_boyas.iterrows():   
    start = dt.datetime.strptime('01-01-1979', '%d-%m-%Y')
    finish = dt.datetime.strptime('31-03-2019', '%d-%m-%Y')
    ax.barh(df_hindcast.loc[i]['Nombre'], (finish - start).days, left=start, color='blue', edgecolor='black')
    
    start = dt.datetime.strptime(task['InicioMedida'], '%d-%m-%Y')
    finish = dt.datetime.strptime(task['UltimaMedida'], '%d-%m-%Y')
    ax.barh(task['Nombre'], (finish - start).days, left=start, color='skyblue', edgecolor='black', alpha=0.8)
ax.set_xlabel('Tiempo')
ax.set_ylabel('Boyas')
ax.set_title('Mediciones')
plt.gca().set_facecolor("white")
ax.grid(True)

#### Explorar datos

data = xr.open_dataset(f'data/raw/historico/GOW_CFS_Bilbao-Vizcaya.nc')
df_var = pd.DataFrame(columns=['Variable', 'Unidades', 'Descripción'])
for var in data.variables:
    standard_name = data[var].standard_name if hasattr(data[var], 'standard_name') else '****'
    units = data[var].units if hasattr(data[var], 'units') else '****'
    df_var.loc[len(df_var.index)] = [var, units, standard_name]
df_var
