==> 08_Analyze_model.ipynb <==
import pandas as pd
pd.options.display.float_format = '{:.4f}'.format

df = pd.read_csv('out_model_red.csv')

# Ordenar por 'loc', 'model' y 'mse'
df = df.sort_values(by=['loc', 'model', 'mse'], ascending=[True, True, True])
# df
# Crear la columna de ranking
df['rank'] = df.groupby(['loc', 'model'])['mse'].rank(method='first')

# Asegurarte de que los rankings sean enteros
df['rank'] = df['rank'].astype(int)

df_grouped = df.groupby(['optimizer', 'n_layers', 'n_neurons']).agg(
    learning_rate=('learning_rate', 'min'),
    ranking=('rank', 'sum')
).reset_index()
df_grouped.sort_values(by=['ranking'], ascending=[True])

df = pd.read_csv('out_model_kmeans.csv')

# Ordenar por 'loc', 'model' y 'mse'
df = df.sort_values(by=['loc', 'model', 'mse'], ascending=[True, True, True])

# Crear la columna de ranking
df['rank'] = df.groupby(['loc', 'model'])['mse'].rank(method='first')

# Asegurarte de que los rankings sean enteros
df['rank'] = df['rank'].astype(int)

df_grouped = df.groupby(['k']).agg(
    ranking=('rank', 'sum')
).reset_index()
df_grouped.sort_values(by=['ranking'], ascending=[True])
