==> 08_Analyze_stats.ipynb <==
import pandas as pd
import numpy as np
pd.options.display.float_format = '{:.3f}'.format

df_res = pd.read_csv('res.csv')

models = df_res['Modelo'].unique()
n_models = len(models)

bias_gow = np.array([df_res[df_res['Modelo'] == model]['bias_gow'] for model in models])
bias_gow.T

bias_cop = np.array([df_res[df_res['Modelo'] == model]['bias_cop'] for model in models])
bias_cop.T

rmse_gow = np.array([df_res[df_res['Modelo'] == model]['rmse_gow'] for model in models])
rmse_gow.T

rmse_cop = np.array([df_res[df_res['Modelo'] == model]['rmse_cop'] for model in models])
rmse_cop.T

p_gow = np.array([df_res[df_res['Modelo'] == model]['pearson_gow'] for model in models])
p_gow.T

p_cop = np.array([df_res[df_res['Modelo'] == model]['pearson_cop'] for model in models])
p_cop.T

si_gow = np.array([df_res[df_res['Modelo'] == model]['si_gow'] for model in models])
si_gow.T

si_cop = np.array([df_res[df_res['Modelo'] == model]['si_cop'] for model in models])
si_cop.T

print((bias_gow.mean(axis=1)+bias_cop.mean(axis=1))/2)
print((bias_gow.std(axis=1)+bias_cop.std(axis=1))/2)

print((rmse_gow.mean(axis=1)+rmse_cop.mean(axis=1))/2)
print((rmse_gow.std(axis=1)+rmse_cop.std(axis=1))/2)

print((p_gow.mean(axis=1)+p_cop.mean(axis=1))/2)
print((p_gow.std(axis=1)+p_cop.std(axis=1))/2)

print((si_gow.mean(axis=1)+si_cop.mean(axis=1))/2)
print((si_gow.std(axis=1)+si_cop.std(axis=1))/2)
