==> 08_View_Stats.ipynb <==
import pandas as pd
pd.options.display.float_format = '{:.4f}'.format
import numpy as np

df_res = pd.read_csv('res.csv')
df_res[df_res['Modelo'] == 'Red_Neuronal']

for column in df_res:
    if column != 'Nombre' and column != 'Modelo':
        print(f"{column}: {df_res[df_res['Modelo'] == 'Red_Neuronal'][column].mean():.4f}")

df_res = pd.read_csv('out_model_red.csv')
print(f"min mse:{min(df_res['mse'])}. max mse: {max(df_res['mse'])}")

# df_res = df_res[df_res['mse'] < 0.5]
# print(f"min mse:{min(df_res['mse'])}. max mse: {max(df_res['mse'])}")

df_grouped = df_res.groupby(['optimizer', 'learning_rate', 'n_layers', 'n_neurons']).agg(
    mse_mean=('mse', 'mean'),
    mse_min=('mse', 'min'),
    mse_max=('mse', 'max')
).reset_index()
df_grouped.sort_values(by=['mse_mean', 'mse_max'], ascending=[True, True])
